requirements.txt
imap-tools
python-dotenv
pandas